+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * TableScrollbarThemeData.verticalOnlyWhenNeeded has been added
   * HeaderThemeData.visible has been added
   * CellIcon.background and CellIcon.alignment removed (already exists on CellStyle)
+  * EasyTable.cellContentHeight moved to CellThemeData.contentHeight
 
 ## 1.6.0